requirements.txt
streamlit==1.25.0
pdf2image==1.16.3
Pillow==9.2.0
poppler-utils==0.86.1